refactor: 카탈로그 조회 limit 추가 #27

+++ src/main/java/sogang/capstone/editking/application/form/FormCatalogFacade.java
@@ -12,7 +12,7 @@ public class FormCatalogFacade {
 
     private final FormCatalogService formCatalogService;
 
-    public FormInfo.CatalogMain retrieveFormCatalog(User user, String status, Long limit) {
+    public FormInfo.CatalogMain retrieveFormCatalog(User user, String status, Integer limit) {
         return formCatalogService.retrieveFormCatalog(user, status, limit);
     }
 }
